新增mapper层，然后在test里面测试napper的分页功能

+++ xuecheng-plus-base/src/main/java/com/xuecheng/base/model/PageResult.java
@@ -20,10 +20,10 @@ public class PageResult<T> implements Serializable {
 
     private List<T> items;
 
-    private int count;
+    private long count;
 
-    private int page;
+    private long page;
 
-    private int pageSize;
+    private long pageSize;
 
 }
